// src/services/messageHandler/messageHandler.types.ts
export enum ACTION {
  IS_CONNECTED = "isConnected",
  REQ_TO_CONNECT = "reqToConnect",
  REQ_TO_SIGN = "reqToSign",
  REQ_WALLET_DETAILS = "reqWalletDetails",

  AUTH_STATUS = "authStatus",

  ERROR = "error",
  CONNECTION_STATUS = "connectionStatus",
  WALLET_DETAILS = "walletDetails",
  SIGNATURE = "signature",
}